gtk3 a11y: Report object's locale
This is basically the gtk3 version of

    Change-Id: I33b3bb406da5d84c2353bbeb5bedc7ac0297f31f
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed May 28 22:54:57 2025 +0200

        qt a11y: Report locale for Qt >= 6.10

and causes the paragraph locales to be reported for the
use case described in the above-mentioned commit also
when using the gtk3 VCL plugin:

    In [14]: acc.get_object_locale()
    Out[14]: 'en_GB.UTF-8'
    In [15]: acc.get_object_locale()
    Out[15]: 'zh_CN.UTF-8'

Add the character encoding at the end of the locale string,
as this is consistent with what is reported by default for
the object locale (was "en_GB.UTF-8" for any object prior
to this change) for the gtk3/ATK case.

The atk_object_get_object_locale doc [1] says that the
returned string is owned by the instance, so make sure to
free the string again by setting it as data for the
AtkObjectWrapper object and free it in
atk_object_wrapper_finalize.

[1] https://docs.gtk.org/atk/method.Object.get_object_locale.html

Change-Id: Icaadc5228f2cc5526b427cc3aef127a8eadd94ff
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185982
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>