tdf#130857 svx: Keep current search term in find dialog
In a WIP branch declaring support for using native Qt
widgets for the "Tools" -> "Find and Replace" dialog,
the scenario

1) open the dialog using "Tools" -> "Find and Replace"
2) enter "foo" in the "Find" combobox
3) press "Replace All" button
4) enter "bar" in the "Find combobox"
5) press "Replace All" button
6) repeat steps 2-5 a few times

would sometimes result in the current search term
("foo" or "bar") getting replaced with the previous
one after pressing the "Replace All" button.

This is because SvxSearchDialog::Remember_Impl
was removing the entry for the current search
term from the combobox and then reinserting it
at index 0.

That would however not result in the "new" entry
to automatically get selected for the Qt implementation
in QtInstanceComboBox, while it seems to be the case
for the GTK and VCL ones.

Avoid the problem by skipping the removal and
reinsertion if the current search term is already
the first one among the combobox entries and in the
vector of remembered search/replace terms.

Change-Id: Ic1d002d414d5b75beb7a80e9ae7125e116a246f4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185987
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>