tdf#166319 sw interdependent redlines: handle accept of del-then-fmt's fmt
The bugdoc has <del>AA<format>BB</format>CC</del> in it, accepting the
BB part resulted in <del>AA</del>BB<del>CC</del>, while the expected
result would be removal of BB from the document.

The problem is that reject only saw the 'format' in the combined
delete+format redline, but that's not correct: it's just a small detail
that there was a format on top of the delete.

Fix the problem by extending
sw::DocumentRedlineManager::AcceptRedlineRange() to handle the
delete-then-format case as a delete and only then do the generic redline
handling.

The surrounding plain delete redlines are not combined yet.

Change-Id: I7ea92a86d8881694ad3da571831100f5996b5d3f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185984
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>