svx: Simplify SvxSearchDialog::Remember_Impl use
So far, the method was called with either

* the current text of the search combobox and a value of bSearch=true
* the current text of the replace combobox and a value of bSearch=false

The method already uses the corresponding combox depending
on the value of the `bSearch` param.

Therefore move the logic to retrieve the current text also in there
and drop the first method param.

The checks whether the current text was empty before calling
the method was also superfluous, because the method returns
early in that case anyway.

Change-Id: Ieb3f20f36ff0bbc7bc33b4f7676ce8df3d20c83f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185989
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins