tdf#130857 qt weld: Prepare for more conditional UI file support
Various .ui files are used for non-toplevel widgets
that can be used from different dialogs/toplevels.

Using native Qt widgets for these is only supported
in case the whole dialog is a native Qt dialog,
because proper functioning depends on a QWidget
hierarchy.
A SalInstanceWidget parent doesn't support
QtInstanceWidget children but relies on them being
SalInstanceWidgets as well (i.e. vcl::Window based).

Prepare for supporting more .ui files conditionally,
i.e. only when the parent is a QWidget as well.

This generalizes the solution introduced with

    commit 9a03fb9266440013eef5f8cd11cc1e5478706496
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Mon May 12 14:05:23 2025 +0200

        tdf#130857 qt weld: Support new Welcome dialog

a bit to use a set of .ui files as well, so more
can easily be added in the future.

Change-Id: I0174d306187ed65f49f165b9eff7f2f2ea05a3a5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186017
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins