tdf#130857 qt weld: Support sc dialog to group by date
Declare support for the Calc "Grouping" dialog to group
by date.
This means that native Qt widgets are used for that dialog
now when using the qt5 or qt6 VCL plugin and starting LO with
environment variable SAL_VCL_QT_USE_WELDED_WIDGETS=1 set.

The dialog can be triggered as follows:

1) Open attachment 184921 from tdf#153217 in Calc
2) switch to "Pivot_Table_Sheet1_1" spreadsheet
3) select cells A1:A13 (i.e. the month names)
4) "Tools" -> "Group and Outline" -> "Group"

This makes use of QtInstanceCalender introduced
in previous commit

    Change-Id: I8a65701a1aa477a04dcf66f63df7864db0b0cdf4
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu May 29 15:32:54 2025 +0200

        tdf#130857 qt weld: Introduce QtInstanceCalendar

and the conditional support for .ui files from

    Change-Id: I0174d306187ed65f49f165b9eff7f2f2ea05a3a5
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu May 29 15:43:15 2025 +0200

        tdf#130857 qt weld: Prepare for more conditional UI file support

Support for svt/ui/datewindow.ui which contains the
the popup that contains the "GtkCalendar" is conditional
on the parent being a native QWidget.

This makes it possible for this to be used for the
newly supported dialog without breaking dialogs
that are not native Qt dialogs (yet), e.g. in
"File" -> "Properties", the "Custom Properties"
tab still uses the vcl::Window based weld::Calendar.
Sample use:

* "File" -> "Properties"
* switch to "Custom Properties" tab
* press "Add Property"
* for the newly created entry, select type "Date"
* click on the menu button in the "Value" column
  to open the calendar popup

(This wouldn't work with the native Qt
version as of now, because the SalInstanceMenuButton
expects to also have a vcl popup.)

Change-Id: I57f4a864651e32a1cfc30f2778d86cb1e89f3349
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186019
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>