tdf#166319 sw interdependent redlines: combine on accept of del-then-fmt's fmt
The bugdoc has <del>AA<format>BB</format>CC</del> in it, accepting the
BB part resulted in <del>AA</del><del>CC</del>. The removal of BB is
good, but there was no combine to also remove AA and CC.

The problem is that the delete-then-format wasn't combined with the
surrounding delete redlines.

Fix the problem by introducing a new CanReverseCombineTypesForAccept()
for the case when the redline with 2 types is conditionally combined
with an other redline and by using that in
sw::DocumentRedlineManager::AcceptRedlineRange().

Also extract some duplicated code into a new lcl_AcceptInnerDelete().

Change-Id: Ib5881b6c9aa44ae5dcf79342dbd64bbd61f734c2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186029
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>