cool#12147: Revive previous copy behaviour on load.
We added some improvements for copying the comments along with their replies before. Which is done with commit:
* https://gerrit.libreoffice.org/c/core/+/177396

But it somehow broke docx export, when there are comments in file.

TextNodes are copied while loading the document. And the error seems to occur in this phase.
This commit keeps the previous behaviour while loading the file.

Signed-off-by: Gökay Şatır <gokaysatir@gmail.com>
Change-Id: I80c343cbc4886f4277e38ff23553195cc8b1acfb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186043
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186605
Tested-by: Jenkins
(cherry picked from commit e2d32fec07396a75754a8979db6108213df11114)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186620
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>