tdf#166319 sw interdependent redlines: handle reject of del-then-fmt's fmt
The bugdoc has <del>AA<format>BB</format>CC</del> in it, rejecting the
BB part is meant to get rid of the delete redline and keep the format
redline.

The problem is that the old code only considered the primary type
(format) of the redline, but that's not correct: it's just a small
detail that there was a format on top of the delete.

Fix the problem by extending
sw::DocumentRedlineManager::RejectRedlineRange(), so it ignores format
for both insert as delete.

Also simplify sw::DocumentRedlineManager::AcceptRedlineRange(), we have
3 non-overlapping cases there, so no need for bHandled there.

Change-Id: I2607740dd0c1ddc447bb3f761f06c07fe80486b9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186123
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>