tdf#166795 sw: fix removing indent of para with num rule which isn't counted
Go to the end of the bugdoc, press backspace to make the paragraph not
counted in the bullet list, then press backspace again. The cursor
should move to the left, but it stays unchanged.

This went wrong in commit db115bec9254417ef7a3faf687478fe5424ab378
(tdf#78510 sw,cui: split SvxLRSpaceItem for SwTextNode,
SwTextFormatColl, 2023-02-24). The split of SvxLRSpaceItem to 3 separate
pool items is great when it comes to modeling style inheritance
correctly, but this broke reducing the indent on backspace in the bugdoc
case.

Fix the problem by extending the range of pool items we fetch at the
cursor position in SwWrtShell::TryRemoveIndent(), once both the first
line and the text left margin is fetched, this starts to work again.

Note that Word turns off the numbering on the first backspace, while
Writer makes the paragraph not counted instead -- this difference is not
changed here.

Change-Id: I9ac07a99befd59a54b566e61ff1c00277942b52e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186130
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Signed-off-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186132
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>