tdf#164970 [API CHANGE] simplify theme schema
when the feature (LibreOffice Themes) was released, there was a themes
dropdown from where the user could change the theme, and the appearance
radio buttons group which changed the appearance mode of the application

to remain consistant with the two ways of changing application
appearance, a colorscheme schema had Light and Dark entries for Light
and Dark colors for the same entry so that when the user changes
appearance radio buttons from Light to Dark, we could switch the colors.

but that was really confusing, so as per the request tdf#164970, the ui
was redesigned, and the appearance radio buttons were merged with the
themes dropdown (see https://gerrit.libreoffice.org/c/core/+/184429).

this patch cleans up the "Light and Dark" entries from theme schema
and adds just a single entry per color "Color". one theme should have
only one palette of colors, to keep things simple.

Change-Id: Ic832169fb07c2a7e95bbb02e894ecd59a8ac4cbd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184545
Tested-by: Jenkins
Reviewed-by: Sahil Gautam <sahil.gautam.extern@allotropia.de>
Reviewed-by: Vernon, Stuart Foote <vsfoote@libreoffice.org>