Drop an exclusion that seems obsolete
During a debug session, I saw these warnings appearing all the time:

> <SwFlowFrame::MovedBwd(..)> - incorrect next found.

It turned out, that the code emitting them found tables, which were
not in OSL_ENSURE's condition. It didnэt warn about tables prior to
commit 960ef1a29fe441360f6b9536f4239656566f06cf (INTEGRATION: CWS
swqbugfixes11 (1.43.40); FILE MERGED, 2005-01-05); the said commit
added a condition to enter the search loop and eventual "is locked"
check, which excluded the case when the initially found frame was
a table. But it didn't handle the case of a table found in the loop.

Adding such a check looks unreasonable now, when the code didn't
need better filtering of tables in all these years. Before adding
the table back to the condition of OSL_ENSURE (to silence the wrong
warning), I checked if the whole change is indeed still needed. It
turns out, that reverting the change completely does not break the
bugdoc in i#38232, nor any unit tests. So it looks like this may be
reverted as obsolete.

Unfortunately, the bugdoc uses fonts that are unavailable in unit
tests, and it's impossible to re-create the bug to prepare another
test document, so no unit test.

Change-Id: I3d451dd6ebd0d54fbef2f16c1406f7c2693e7cab
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186235
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins