tdf#166850 Apply converted style name to StyleRef field
As a result of e45f187870362ed4448f4a9fb96883b8431887fb the StyleRef
field needs to use the modified style name with (suffix " (user)")
The related JUnit test requires a German locale to detect the
related issue. Otherwise there is no suffix that needs to be
applied to the field.

Change-Id: Iad4257568ad05900df63a6098ba81e926df701e1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186150
Reviewed-by: Thorsten Behrens <thorsten.behrens@allotropia.de>
Tested-by: Jenkins