Related: tdf#166208 Avoid C-style array and fixed index madness
Here we deal with conditional length and set of elements in the two
sequences (names, values).  This required a complex logic to handle
them; and all the changes required accurate synchronization both in
the code, and with the schema.  An example of a confusion from this
is commit 19f3b72f34c487dc97d582712d21734a7e055fd5, which increased
the size of the sequence returned from lcl_CreatePropertiesNames.

Instead, use ConfigItem::GetNodeNames to get the defined names from
schema.  This unifies the code for all eCall cases, by only reading
and writing those values that exist in the respective config node's
schema.

Change-Id: I90fdaa603bb12dc554a4c2947dd7f3415af31f24
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186236
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>