Related: tdf#166208 Drop bBeforeDetection
It was intended to detect the first use of the text import dialog
after introduction of the automatic encoding detection feature.
But since the schema for SeparatorType property, which was used
for the detection, defined it as non-nillable, the property has
a value always, so the detection is impossible. It means that the
detection code may be removed.

The detection, if it worked, would prevent tdf#166208; it would
select automatic encoding detection, and prevent the shift of the
old selection by one.

Change-Id: I1ba661476531987633e0ed0a100bf798aa01a9f6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186239
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
(cherry picked from commit 58dc72c6f29a3ad4cbea09803d6c568b96f80c08)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186245
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>