tdf#166871: drop mnMaxParaPerPage hack
It was there since commit 84a3db80b4fd66c6854b3135b5f69b61fd828e62
(initial import, 2000-09-18); and its idea was to use some heuristics
to predict the likely number of pages (when there was no metadata, to
use that number for progress indicator), and, most importantly, to
predict the page breaks (where a paragraph must be moved to the next
page, just because there are many paragraphs on this page already; it
likely was considered an optimization).

But that hack, in fact, creates page breaks in random places, and
then requires moving stuff back; in some cases, the resulting layout
was instable, and layout loop control triggered. And that broke the
most straightforward way of arranging pages, from start to end.

This change removes that hack. The speedup that the hack could
potentially provide would be negligible, if at all; and the damage
is real.

I don't know if the following part of documentation comment for the
constructor of SwLayHelper is still relevant after this change:

 "If there's no layout cache, the distribution to the pages is more
  a guess, but a guess with statistical background."

testObjectCrossReference needed a fields update, for unclear reason.

SwTabFrame::Split needed to be improvedto fix failing TestTdf150606,
where a zero-height column body tried to layout its table; in this
case, in SwTabFrame::Split pRow was pointing to Lower(), nRowCount
was 0, nRepeat was 0, and bSplitRowAllowed was false -> it chose
"bKeepNextRow" path, moved pRow to its next, split it, and a wrong
single-cell table appeared on page 3. This seems to pre-exist, and
only happened to be manifested in a test by the change.

Change-Id: I2d203055e67b0155bcd719efb1f8a8b6a82b1156
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186243
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins
Signed-off-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186289