tdf#166319 sw interdependent redlines: combine on reject of del-then-fmt's fmt
The bugdoc has <del>AA<format>BB</format>CC</del> in it, rejecting the
redline under the cursor when it's inside BB resulted in rejecting BB,
but not AA or CC.

The problem is that the delete-then-format wasn't combined with the
surrounding delete redlines.

Fix the problem by extending the reverse combine code in
sw::DocumentRedlineManager::RejectRedlineRange() to recognize
delete-then-format when it recognized insert-then-format already.

Also rename CanReverseCombineTypesForAccept() to communicate this is now
used for reject purposes, too.

Change-Id: I7832af605d09ce9a29dd10d68c1915e7f7becfbd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186215
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>