tdf#165742 Step 4.4: Establish a narrow export path for chartex
This is a subtask of tdf#165742: Chartex charts are lost on input from
OOXML and re-export.

Fix chartex parsing so as to properly handle series. This code considers
the <cx:plotAreaRegion> tag to be equivalent to the chart 'type group'
tags (<c:barChart>, <c:areaChart>, etc.). Therefore it allows use of the
type group mechanisms in chartex, even though there is not strictly a
tag for type group. This also includes a basic CI test for the
ability to round-trip (OOXML -> LO -> OOXML) simple chartex content.

Change-Id: Ibe96a239aa5329897b8e0652ce7388a7c398dda2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186718
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>