tdf#166940 sw: in docx field import use to outer char props
For RTF field import, to assign character properties top a field, the
properties are assigned to the field mark characters and then added to
the content string.

For DOCX according to the standard the content can have arbitrary
styles, but for now we only support plain strings for some fields. So,
for those fields, we want to insert the content with the properties of
the surrounding context. This was done by taking the character
properties of the end field mark character, like used in RTF import.
This works most of the time but fails for documents where those
invisible characters have their own properties which shouldn't show up
in the visible text.

Instead, for DOCX, it makes more sense to take those properties from the
first run of the field and apply it to the whole field result.

Change-Id: I9287f7ed81c4cc2b5bbf65e7151ee3a8bcbd95d2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186302
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
(cherry picked from commit adc483d5f9256b7e10beed29bb63a7c582cab56e)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186453
Tested-by: Jenkins
Reviewed-by: Jaume Pujantell <jaume.pujantell@collabora.com>