tdf#166879 RTF export: handle section break right after a section
The RTF file opens fine, but saving and opening in either Writer or Word
loses the section break between the two paragraphs.

What happens is that the first paragraph is in a Writer section, and the
second paragraph has an SwFormatPageDesc associated with it. And
RtfExport::OutputEndNode() only emits section breaks for end nodes if
they are the end of a table.

Fix the problem by also handling the end of Writer sections, similar to
how it already works for the DOCX export since commit
7060c7b642fdc0a369505e430652ee44205e7eed (tdf#95367 DOCX: allow r-t of
changed first/follow sections, 2016-10-10).

The original scenario was around saving a file for the 2nd time, the
testcase includes a subset of Writer's own RTF export result, so just
one import + export is needed.

Change-Id: I5cbb9d1ac740158f798fa43c4357c23d8037ba27
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186317
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>