Related: tdf#163008 draw the selected tab using a push button on macOS
On macOS, more closely match the vertical tab style of the
sidebar in the System Settings application. As of macOS Sequoia,
the sidebar only shows the default style push button for the
selected tab. The unselected tabs are only text and images so
no native control needs to be drawn for unselected tabs.

Also, remove the vertical tab's 3D border.

Change-Id: I0df77b7613e37897213ce656f894f324ebf2ba14
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186340
Reviewed-by: Heiko Tietze <heiko.tietze@documentfoundation.org>
Reviewed-by: Patrick Luby <guibomacdev@gmail.com>
Tested-by: Jenkins