tdf#166943: zero passed to AllowMoreInput means "no more input"
Users of ToggleUnicodeCodepoint keep reading next characters to the left,
until AllowMoreInput returns false (or selection ends). Attempting to
read more characters to the left then there are, SwCursorShell::GetChar
returns 0. This needs to be treated by AllowMoreInput as hard stop. It
failed, when the previous character was a combining character.

Change-Id: I203b150154e1948d4cebfd69442e30a076710f46
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186341
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>