Drop obsolete asserts
These were added in commit d21298e3b108b0dfdb1515248068314757cf4dd7
(coverity#1438401 assert nPrevIndex >= 0, 2018-08-18), and commit
3e97b8a74b20d72d36ac65d3ef091cae79b8cdac (coverity#1438402 Improper
use of negative value, 2018-08-20); the description of the problem
in Coverity issue was "The negative value may be unexpected by later
operations, causing incorrect computations".

The index returned from GetStyleNameIndex indeed can be negative; it
is passed to OpenAndCloseRow or directly to OpenNewRow, and eventually
to WriteRowStartTag, which was fixed to treat negative index in commit
7246759822aff30e4e4e1bca7949eae3b0d960ef (tdf#143940: the style index
might be -1, 2021-08-27).

Change-Id: Ia0d737d61e5cd31d964b77facd09b4fb79e34e4f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186347
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins