tdf#166428 accept stock charts with only low and high series
A stock chart may consist of only series 'Low' and 'High' in UI. But
the method MergeSeriesForStockChart() expects three series. The third
one is for a series 'Close'. That corresponds to ODF. For this third
series, LO writes a series without chart:values-cell-range-address. That
is allowed in ODF.

mbAllRangeAddressesAvailable is set to 'false' on import, because of
this dummy series. Thus MergeSeriesForStockChart() was not executed.
LO 24.8 has used as fallback, to generate the chart from the range
stored in msChartAddress. This member gets its value from attribute
table:cell-range-address of the chart:plot-area element.

This attribute was already deprecated for long time and is now removed
in ODF 1.4. Therefore the fallback does not work any longer for documents
saved in ODF 1.4, which is the default for LO 25.x. Loading such chart in
LO 25.x had lost all series.

Fortunately it is not necessary to skip MergeSeriesForStockChart() if
there is an empty series. The method can already handle it itself. The
fallback would have not been necessary in LO 24.8.

The patch changes the procedure so, that MergeSeriesForStockChart is
called even if mbAllRangeAddressesAvailable is false. I don't know
whether the fallback is used for other chart types in documents
producted by older LO or OOo versions. Thus I have not changed the
else-branch.

Change-Id: I2f059cb1778cf84161cf08b35c27ef50a66aefe3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186386
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>
Tested-by: Jenkins
(cherry picked from commit af99860d42c672c3c47aab5e3b55f0bb52df4b4e)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186410
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>