tdf#166319 sw interdependent redlines: add UI to create format inside insert
The bugdoc has <ins>AABBCC</ins> in it, selecting BB and pressing e.g.
bold simply applied the formatting without creating a redline, even if
the change is from another user.

This was working for the case when a delete redline is created on top of
an insert one, but was broken for insert-then-format: simply code was
missing to try to handle the non-IsOwnRedline() case.

Fix the problem by extracting this "split existing redline as necessary,
then turn the overlapping range into a hierarchical redline" logic from
sw::DocumentRedlineManager::PreAppendDeleteRedline() into a new
PreAppendForeignRedline(), and next to the old insert-then-delete case,
also use it for insert-then-format and delete-then-format.

The new PreAppendForeignRedline() has a single part where it deals with
moves for insert-then-delete, disable that for insert-then-format, since
that's not relevant. Undo still needs to be fixed. Finally keep the
logic unchanged the own delete/insert + own format redline case for now,
we want different behavior for insert-then-delete and insert-then-format
in this case, so intentionally not sharing that.

Change-Id: Ifd2903f39d92a1bee96de0a38976eb5111c49223
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186423
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins
(cherry picked from commit eef0dfed817e40cd83e8ba8e290f45c224257f97)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186440