tdf#166319 sw interdependent redlines: undo of creating an ins-then-fmt redline
The bugdoc has <ins>AABBCC</ins> in it, selecting BB, marking the
selection as e.g. bold + undo resulted in <ins>AA</ins>BB<ins>CC</ins>
instead of the original <ins>AABBCC</ins>.

What happened is that the created SwUndoAttr didn't record the deleted
insert redline, so there was no chance to restore it. The same was
working for the insert-then-delete case via SwUndoRedlineDelete, because
that stored the direct delete redline as an SwRedlineData, and also
stored the underlying insert redline inside SwRedlineSaveDatas.

Fix the problem by extending the types of redlines we save in
SwUndo::FillSaveDataForFormat(), so SwUndoAttr also saves the insert
redline in its SwRedlineSaveDatas array, then undo starts to work.

Redo still needs fixing, though.

Change-Id: Ic436974f15633eabcc298f062e57183220a48b27
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186449
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>